+++ mne/io/kit/coreg.py
@@ -22,6 +22,7 @@ from ...transforms import (
     Transform,
     als_ras_trans,
     apply_trans,
+    fit_matched_points,
     get_ras_to_neuromag_trans,
 )
 from ...utils import _check_fname, _check_option, warn
@@ -133,7 +134,7 @@ def _set_dig_kit(mrk, elp, hsp, eeg, *, bad_coils=()):
     hpi_results : list
         The hpi results.
     """
-    from ...coreg import _decimate_points, fit_matched_points
+    from ...coreg import _decimate_points
 
     if isinstance(hsp, str | Path | PathLike):
         hsp = _read_dig_kit(hsp)
